chore: refactor Docker command handling by consolidating container checks into ChecksDocker trait
